+++ MeDrop/MeDrop/Sources/MeDropApp.swift
@@ -13,8 +13,7 @@ struct MeDropApp: App {
       @UIApplicationDelegateAdaptor(AppDelegate.self) var dele        
     var body: some Scene {
         WindowGroup {
-            ShareView()
-//             MainView()
+             MainView()
                 .preferredColorScheme(.light)
 //                .onAppear{
 //                    loadUserData()
